chain: move LockTime into transaction
